utils for static in own files
