+++ pages/api/chat.ts
@@ -46,80 +46,12 @@ const handler = async (req: Request): Promise<Response> => {
 
     encoding.free();
 
-    console.log('>>> on chat')
-    console.log(model, promptToSend, key, messagesToSend)
     const stream = await OpenAIStream(model, promptToSend, key, messagesToSend);
 
     return new Response(stream);
   } catch (error) {
     console.error(error);
     return new Response('Error', { status: 500 });
-  }
-};
-
-export const apiChat = async (options) => {
-  try {
-    const { model, messages, key, prompt } = options as ChatBody;
-
-console.log('>>> messages', messages)
-    //await init((imports) => WebAssembly.instantiate(wasm, imports));
-    /*
-    const encoding = new Tiktoken(
-      tiktokenModel.bpe_ranks,
-      tiktokenModel.special_tokens,
-      tiktokenModel.pat_str,
-    );
-
-    const tokenLimit = model.id === OpenAIModelID.GPT_4 ? 6000 : 3000;
-*/
-    let promptToSend = prompt;
-    if (!promptToSend) {
-      promptToSend = DEFAULT_SYSTEM_PROMPT;
-    }
-/*
-    const prompt_tokens = encoding.encode(promptToSend);
-
-    let tokenCount = prompt_tokens.length;
-    let messagesToSend: Message[] = [];
-
-    for (let i = messages.length - 1; i >= 0; i--) {
-      const message = messages[i];
-      const tokens = encoding.encode(message.content);
-
-      if (tokenCount + tokens.length > tokenLimit) {
-        break;
-      }
-      tokenCount += tokens.length;
-      messagesToSend = [message, ...messagesToSend];
-    }
-
-    encoding.free();
-*/
-
-    let messagesToSend: Message[] = [];
-    for (let i = messages.length - 1; i >= 0; i--) {
-    
-      const message = messages[i];
-      /*
-      // @to-do check limit
-      const tokens = encoding.encode(message.content);
-
-      if (tokenCount + tokens.length > tokenLimit) {
-        break;
-      }
-      tokenCount += tokens.length;
-      */
-      messagesToSend = [message, ...messagesToSend];
-    }
-    console.log('>>> on chat')
-    console.log(model, promptToSend, key, messagesToSend)
-    const stream = await OpenAIStream(model, promptToSend, key, messagesToSend);
-    console.log('>>> stream', stream)
-    //return new Response(stream)
-    return stream
-  } catch (error) {
-    console.error(error)
-    return { error }
   }
 };
